Changed setupEventListener to add the event listener immediately.
Previously it would wait to add the internal event listener until after the Google object was ready. This could cause issues with other code that depending on the event getting added immediately becuase it used hasListener(). Because there could be a delay in adding the event while the Google library is loaded, hasListener() would incorrectly return false even though the event was set.

Now, the only pending event handling is when the Google library hasn’t loaded yet or the Google object isn’t set up yet. In those cases the event will be added to the Google object once it’s set.