Undid saving the tooltip for later
Instead, when the click or hover event occurs, get the map then. This way the map value isn’t dependent on what it is when the tooltip is attached.

This lets the tooltip get attached to a marker that isn’t assigned to the map yet. The map value is retrieved on demand when the click/hover event happens.