Fixed bug where the latLngObject was linked to the property and updating weird
The latLngObject was assigned to the northEast and southWest properties. But when extend() was called again the latLngObject value was linked to northEast and southWest so the behavior was unexpected. The end result was that the northEast and southWest values would take on the last latLngObject value.
The solution was to clone the object when assigning it to a property.