Fixed issue where the marker anchorPoint is sometimes not defined
It’s possible that when a marker is first rendered that it’s anchorPoint has not been set yet. This will try with setTimeout up to 5 times to get the anchorPoint.
This shoudn’t ever happen when the popup is shown when clicking on the marker. It may only happen when programatically showing the popup when the map and/or marker is first rendered.