Don’t set up the marker object multiple times when setting up options
Previously, when setting up certain options, the Google Maps marker object would get created. This happened multiple times in the setOptions() method.
Now, each of the options is only set up on the Google Maps marker object if it has already been created. Otherwise, the option value gets set in the class, ready to be added to the Google Maps marker object when it’s created.