Fixed issue where bounds_changed or zoom_changed wouldn't be called
This happens when the events are set up early in the map loading process. Even though this library did not add those events, the events exist when using google.maps.event.hasListeners().

It's possible that an external library, like the one we use to load the google map, could be adding these events. I suspect this because after the library is loaded the events are no longer set up.