Set the image to be 100% width and height
this is necessary to avoid the image moving on the map when zooming in. If the image doesn’t naturally fill the “box” created by the latitude/longitude bounds and it’s not set to 100% width and height, then when zooming in, the image may shift and move on the map.